deleted my PageHeaderAdminDropdownMenuTest class moved my testBuildsOpenFromDropdownMenu to Headertest class
