+++ src/test/java/school/redrover/HeaderTest.java
@@ -1,6 +1,7 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
@@ -239,5 +240,22 @@ public class HeaderTest extends BaseTest {
                     "The result list doesn't contain the item " + ITEM_NAME);
         }
         softAssert.assertAll();
+    }
+
+    @Test
+    public void testBuildsOpenFromDropdownMenu(){
+
+        WebElement dropDownMenu = getWait2().until(ExpectedConditions.presenceOfElementLocated(By.xpath
+                ("//a[@href='/user/admin']//button[@class='jenkins-menu-dropdown-chevron']")));
+        JavascriptExecutor executor = (JavascriptExecutor) getDriver();
+        executor.executeScript("arguments[0].click();", dropDownMenu);
+
+        WebElement btnBuilds = getWait5().until(ExpectedConditions.elementToBeClickable
+                (By.xpath("//div[@id='breadcrumb-menu']//span[.='Builds']")));
+        btnBuilds.click();
+
+        WebElement pageBuilds = getDriver().findElement(By.xpath("//h1[.='Builds for admin']"));
+
+        Assert.assertTrue(pageBuilds.isDisplayed());
     }
 }
